backend-ms1/src/test/java/ma/zs/zyn/unit/ws.facade/admin/forum/ForumStateRestAdminTest.java
package ma.zs.zyn.unit.ws.facade.admin.forum;

import ma.zs.zyn.bean.core.forum.ForumState;
import ma.zs.zyn.service.impl.admin.forum.ForumStateAdminServiceImpl;
import ma.zs.zyn.ws.facade.admin.forum.ForumStateRestAdmin;
import ma.zs.zyn.ws.converter.forum.ForumStateConverter;
import ma.zs.zyn.ws.dto.forum.ForumStateDto;
import org.aspectj.lang.annotation.Before;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.util.Collections;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class ForumStateRestAdminTest {

    private MockMvc mockMvc;

    @Mock
    private ForumStateAdminServiceImpl service;
    @Mock
    private ForumStateConverter converter;

    @InjectMocks
    private ForumStateRestAdmin controller;

    @Before("")
    public void setUp() {
        MockitoAnnotations.initMocks(this);
        this.mockMvc = MockMvcBuilders.standaloneSetup(controller).build();
    }


    @Test
    public void itShouldFindAllForumStateTest() throws Exception {
        // Mock the service to return an empty list
        when(service.findAll()).thenReturn(Collections.emptyList());
        when(converter.toDto(Collections.emptyList())).thenReturn(Collections.emptyList());

        // Call the controller method
        ResponseEntity<List<ForumStateDto>> result = controller.findAll();

        // Verify the result
        assertEquals(HttpStatus.NO_CONTENT, result.getStatusCode());

        // Response body should be empty list
        List<ForumStateDto> responseBody = result.getBody();
        assertNotNull(responseBody);
        assertEquals(0, responseBody.size());
    }

    @Test
    public void itShouldSaveForumStateTest() throws Exception {
        // Mock data
        ForumStateDto requestDto = new ForumStateDto();
        ForumState entity = new ForumState();
        ForumState saved = new ForumState();
        ForumStateDto savedDto = new ForumStateDto();

        // Mock the converter to return the forumState object when converting from DTO
        when(converter.toItem(requestDto)).thenReturn(entity);

        // Mock the service to return the saved client
        when(service.create(entity)).thenReturn(saved);

        // Mock the converter to return the saved forumState DTO
        when(converter.toDto(saved)).thenReturn(savedDto);

        // Call the controller method
        ResponseEntity<ForumStateDto> result = controller.save(requestDto);

        // Verify the result
        assertEquals(HttpStatus.CREATED, result.getStatusCode());

        // Verify the response body
        ForumStateDto responseBody = result.getBody();
        assertNotNull(responseBody);

        // Add assertions to compare the response body with the saved forumState DTO
        assertEquals(savedDto.getCode(), responseBody.getCode());
        assertEquals(savedDto.getLibelle(), responseBody.getLibelle());
    }

}
